Update PC Tools to VKSC 1.0.6
And continue to support reading version 1.0.4 version of caches.
Support will be removed in the future.
Clients should be updated to avoid the use of the VkPipelineCacheHeaderVersionSafetyCriticalOne
structure directly, and should use the new VKSCPipelineCacheHeaderReader accessor functions
to get the information the need (this enables abstraction of the two different cache header layouts).